+++ CHANGELOG.md
@@ -9,8 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Link attribute `conditional`: to make a link conditional when it has no conditional and is not an error handler.
-- Link attribute `cache_if_not_required`: cache inputs from non-required links just like required links.
+- Added link attribute `cache_if_not_required`: cache inputs from non-required links just like required links.
+
+### Changed
+
+- Handle link attribute value `required=False`: for the link to be explicitly non-required.
 
 ## [4.0.2] - 2026-03-03